reduced number of requests made to MyAnimeList and rating is float
each voice actor made a request to get a higher resolution image. This
isn't necessary because the character/staff endpoint already includes a
picture if the "r/42x62/" part of the URL is being removed

if the rating only had one digit then it was an integer but the metadata
expects a float.