Bugfix for issue #29
When MyAnimeList does not provide an alternative title in English then
that element will be still available in the Jikan API request, but it is
null.

The Agent did warn that the title is not available but didn't update any
metadata. The reason for this is that plex seems to need the Title so
that any other metadata also gets added. The Agent also never fell back
to the main title because the that funcationality wasn't even
implemented.

The Agent now will try to get the title in the language the user has
specified and if that does not work will log a warning in the log files
and fallback to the main title.