Updated Title retrieval from Jikan
Jikan has deprecated the title, title_english, title_japanese and title_synonyms fields in the /anime/id request. They are now an array of strings in the titles element.

this array contains the type (Default, Synonym, Japanese, English) and the Title for that type.

The Setting to retrieve the "main" title is now the Default title to reflect the Default Title element more clearly.

A new Option in the Settings has been added to exclude the Preferred Title language from searching for the Anime Title. This was added to prevent a show suddenly not being matched anymore because the preferred language is different to the default language used for the files and folders