Fix: #38 Instant "No Matches Found" on select shows.
When metadata was requested from a Season that isn't Season 1 the Agent failed because it specifically looked for Season 1 Episode 1 to find the .match file. The Agent will now check the Seasons and output a Warning message for an unsupported Structure (if the Season is not 1) or get the episode number that the request was made with.